requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings
google-generativeai # Mantenemos esta para Gemini
# pypdf ya no es necesaria

# --- Dependencias de Base de Datos ---
sqlalchemy # ORM
psycopg2-binary # Driver PostgreSQL (alternativa: psycopg)
alembic # Para migraciones de base de datos

